LayoutTests/transitions/interrupted-accelerated-transition.html

<!DOCTYPE html>

<html>
<head>
    <style>
        #box {
            display: inline-block;
            width: 100px;
            height: 100px;
            margin: 10px;
            background-color: blue;
            -webkit-transition: -webkit-transform 10s;
        }
        
        #box.rotated {
            -webkit-transform: rotate(45deg);
        }
    </style>
    <script>
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.waitUntilDone();
        }
        function loaded()
        {
            var box = document.getElementById('box');
            box.className = 'rotated box';
            window.location = 'resources/interrupted-accelerated-transition-final.html';
        }
        window.addEventListener('load', loaded, false);
    </script>
</head>
<!-- The unload handler keeps this page out of the page cache. -->
<body onunload="void(0)">
    <div id="box"></div>
</body>
</html>
